move suspense to store-ui
