+++ packages/gatsby-theme-store/package.json
@@ -49,7 +49,6 @@
     "markdown-loader": "^6.0.0",
     "nprogress": "^0.2.0",
     "react-helmet-async": "^1.0.7",
-    "react-hook-inview": "^4.3.0",
     "requestidlecallback-polyfill": "^1.0.2",
     "schema-dts": "^0.6.0",
     "swr": "^0.3.0"
