+++ packages/store-ui/package.json
@@ -34,6 +34,7 @@
     "@vtex-components/accordion": "^0.2.3",
     "@vtex-components/drawer": "^0.2.4",
     "deepmerge": "^4.2.2",
+    "react-hook-inview": "^4.3.9",
     "react-swipeable": "^6.0.0",
     "reakit": "^1.3.0",
     "text-mask-core": "^5.1.2",
